plugin config reorg
